Move _apply_relativistic_doppler_shift to spectral_coordinate.py and move SpectralQuantity to astropy.coordinates
