+++ astropy/units/__init__.py
@@ -35,7 +35,6 @@ from .function.core import *
 from .function.logarithmic import *
 
 from .decorators import *
-from .spectral_quantity import *
 
 del bases
